requirements.txt
aiofiles==23.2.1
altair==5.3.0
altgraph==0.17.4
anyio==4.4.0
asammdf==7.4.5
attrs==24.2.0
blinker==1.8.2
canmatrix==1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi-cli==0.0.4
ffmpy==0.3.2
Flask==3.0.3
fonttools==4.53.1
fsspec==2024.6.1
future==1.0.0
gradio_client==1.0.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib_metadata==8.2.0
importlib_resources==6.4.3
isal==1.7.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.17.3
kiwisolver==1.4.5
lxml==5.3.0
lz4==4.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
numexpr==2.10.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pefile==2023.2.7
pillow==10.4.0
pydub==0.25.1
Pygments==2.18.0
pyinstaller==6.10.0
pyinstaller-hooks-contrib==2024.8
pyparsing==3.1.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.15.0
pyqtgraph==0.13.7
pyrsistent==0.20.0
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32-ctypes==0.2.3
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.32.3
rich==13.7.1
ruff==0.5.1
semantic-version==2.10.0
shellingham==1.5.4
shiboken6==6.7.2
six==1.16.0
sniffio==1.3.1
tomlkit==0.12.0
toolz==0.12.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websockets==11.0.3
Werkzeug==3.0.3
zipp==3.20.0

redis~=5.0.8
openpyxl~=3.1.5
pymysql==1.1.1
DBUtils==3.1.0
docxcompose==1.4.0
cryptography
